+++ static_library/stat_lib.h
@@ -4,5 +4,7 @@
 
 #ifndef DESKTOP_HOMEWORK2_STATIC_LIBRARY_STAT_LIB_H_
 #define DESKTOP_HOMEWORK2_STATIC_LIBRARY_STAT_LIB_H_
+int number_of_sequence(char* chunk, char* sequence);
+int run(char* sequence, int seq_size, char* log_path);
 
 #endif  // DESKTOP_HOMEWORK2_STATIC_LIBRARY_STAT_LIB_H_
